requirements.txt
hydra-core==1.3.2
omegaconf~=2.3.0
torch~=2.5.1
torchinfo==1.8.0
soundfile==0.12.1
pytorch-lightning==2.4.0
tensorboard==2.18.0

torchaudio~=2.5.1
numpy~=2.1.3
tqdm==4.67.0
norbert==0.2.1